+++ Ensure.Web/Data/AppUser.cs
@@ -10,7 +10,7 @@ namespace Ensure.Web.Data
 	public class AppUser : IdentityUser
 	{
 		public List<EnsureLog> Logs { get; set; }
-		public short TimeZone { get; set; } = 2; // UTC+2.0
-		public short DailyTarget { get; set; }
+		public int TimeZone { get; set; } = 2; // UTC+2.0
+		public int DailyTarget { get; set; }
 	}
 }
