+++ Ensure.Web/Services/IAppUsersService.cs
@@ -7,7 +7,7 @@ namespace Ensure.Web.Services
 {
     public interface IAppUsersService
     {
-        public Task SetUserTarget(short target, string userName);
+        public Task SetUserTarget(int target, string userName);
         public Task<int> GetUserTarget(string userName);
         public string GenerateBearerToken(AppUser user);
         public Task<ApiUserInfo> GetUserInfo(string userName, string jwtToken);
